pyproject.toml
[tool.poetry]
name = 'lol_sound_downloader'
version = '0.1.0'
description = ''
authors = ['Victor Otieno Omondi <victor@slick.co.ke>']

[tool.poetry.dependencies]
python = '^3.7'
aiohttp = "^3.6"
pyjwt = "^1.7"
requests = "^2.22"
cryptography = "^2.8"
